Server consuma pacchetti a 30 Hz fissi

+++ Physics/library.cpp
@@ -14,7 +14,7 @@ void uniform_rectilinear_motion(float *position, float velocity, float deltaTime
 
 void normalizeVelocity(float* velX, float* velY) {
     // Calcoliamo la lunghezza del vettore (magnitudo)
-    float length = std::sqrtf(*velX * *velX + *velY * *velY);
+    float length = std::sqrt(*velX * *velX + *velY * *velY);
 
     // Evitiamo la divisione per zero se il pinguino è fermo
     if (length > 0.0f) {
